Refactor - move colors constants to commons to reuse it
