--- app/src/main/java/org/dhis2/usescases/biometrics/ui/Colors.kt
@@ -1,6 +1,0 @@
-package org.dhis2.usescases.biometrics.ui
-
-const val defaultButtonColor = "#007acc"
-const val successButtonColor = "#FF35835D"
-const val failedButtonColor = "#C57704"
-const val declinedButtonColor = "#a6a5a4"
