+++ app/src/main/java/org/dhis2/usescases/biometrics/ui/SequentialNextSearchAction.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import org.dhis2.R
+import org.dhis2.commons.biometrics.defaultButtonColor
 import org.hisp.dhis.mobile.ui.designsystem.component.Button
 import org.hisp.dhis.mobile.ui.designsystem.component.ButtonStyle
