--- form/src/main/java/org/dhis2/form/ui/biometrics/components/Colors.kt
@@ -1,6 +1,0 @@
-package org.dhis2.form.ui.biometrics.components
-
-const val defaultButtonColor = "#007acc"
-const val successButtonColor = "#FF35835D"
-const val failedButtonColor = "#C57704"
-const val declinedButtonColor = "#a6a5a4"
