+++ form/src/main/java/org/dhis2/form/ui/biometrics/components/biometricsTEIRegistration/LinkLastBiometricsCheckBox.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import org.dhis2.commons.biometrics.defaultButtonColor
 import org.dhis2.form.R
-import org.dhis2.form.ui.biometrics.components.defaultButtonColor
 
 @Composable
 internal fun LinkLastBiometricsCheckBox(
