+++ form/src/main/java/org/dhis2/form/ui/biometrics/components/biometricsTEIRegistration/SaveWithoutBiometricsButton.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import org.dhis2.form.ui.biometrics.components.defaultButtonColor
+import org.dhis2.commons.biometrics.defaultButtonColor
 
 @Composable
 internal fun SaveWithoutBiometricsButton(
